requirements.txt
google-adk
google-generativeai
google-cloud-aiplatform
google-cloud-storage
google-cloud-bigquery
google-api-python-client
google-auth
pydantic
PyPDF2
